# architect-go/docs/generated/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Architect Dashboard API",
    "description": "Comprehensive REST API for managing projects, tracking events, and coordinating development workflows",
    "version": "3.2.0",
    "contact": {
      "name": "Support",
      "url": "https://support.architect.example.com",
      "email": "support@architect.example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.architect.example.com",
      "description": "Production"
    },
    {
      "url": "https://staging.architect.example.com",
      "description": "Staging"
    },
    {
      "url": "http://localhost:8080",
      "description": "Development"
    }
  ],
  "paths": {
...
    "/events": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "List Events",
        "description": "Get paginated list of event logs with filtering",
        "operationId": "listEvents",
        "parameters": [
          {
            "name": "event_type",
            "in": "query",
            "description": "Filter by event type",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "source",
            "in": "query",
            "description": "Filter by source",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "Filter by user ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "description": "Filter by project ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page (1-1000)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Items to skip",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event list"
          }
        }
      },
      "post": {
        "tags": [
          "Events"
        ],
        "summary": "Create Event",
        "description": "Create a new event log entry",
        "operationId": "createEvent",
        "requestBody": {
          "description": "Event data",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "object"
                  },
                  "event_type": {
                    "type": "string",
                    "example": "user_action"
                  },
                  "project_id": {
                    "type": "string"
                  },
                  "source": {
                    "type": "string",
                    "example": "api"
                  },
                  "user_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "event_type",
                  "source"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Event created"
          }
        }
      }
    },
    "/events/{event_id}": {
      "get": {
        "tags": [
          "Events"
        ],
        "summary": "Get Event",
        "description": "Get a specific event log entry",
        "operationId": "getEvent",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "description": "Event ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event details"
          },
          "404": {
            "description": "Event not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Events"
        ],
        "summary": "Delete Event",
        "description": "Delete an event log entry",
        "operationId": "deleteEvent",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "description": "Event ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Event deleted"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerToken": {
        "type": "http",
        "description": "Bearer token authentication",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "sessionCookie": {
        "type": "apiKey",
        "description": "Session cookie set after login",
        "name": "session",
        "in": "cookie"
      }
    }
  },
...
}